Fixed some bugs & added log at constructor of EuphonyTx

+++ project/src/main/cpp/EpnyOscillator.cpp
@@ -1,5 +1,5 @@
 //
-// Created by opener on 20. 8. 25.
+// Created by designe on 20. 8. 25.
 //
 
 #include "EpnyOscillator.h"
@@ -20,9 +20,10 @@ void EpnyOscillator::renderAudio(float *data, int32_t numFrames) {
     if(mIsWaveOn) {
         double time = 0.0;
         for(int i = 0; i < numFrames; ++i) {
-            time = i / mSampleRate;
+            time = (double)i / (double)mSampleRate;
             data[i] = sinf(kTwoPi * mFrequency * time) * mAmplitude;
         }
+
     } else {
         memset(data, 0, sizeof(float) * numFrames);
     }
